requirements.txt
asgiref==3.3.1
boto3==1.17.27
botocore==1.20.27
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.7
django-crispy-forms==1.11.1
django-import-export==2.5.0
django-storages==1.11.1
et-xmlfile==1.0.1
jmespath==0.10.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.1.2
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.3
xlrd==2.0.1
xlwt==1.3.0
